automations/webhook-to-google-sheet/requirements.txt
flask>=3.0.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
